langchain-projects/qna_chatbot_ollama/requirements.txt
langchain_ollama
pandas 
numpy
python-dotenv
langchain_core
langchain_community
streamlit